# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion de biens immobilier",
        "description": "Documentation de l'API immoApp",
        "version": "1.0.0"
    },
    "paths": {
...
        "/utilisateurs": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Liste tous les utilisateurs",
                "operationId": "3ca4091dbd6454ea1ec2eb43f5db6c3b",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Utilisateurs"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Créer un utilisateur",
                "operationId": "01edc46161789482f77838a996fd3943",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Utilisateurs"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Utilisateurs"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/utilisateurs/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Afficher un utilisateur",
                "operationId": "df179f2c14f50a5e95965e8bc116aae3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Utilisateurs"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Mettre à jour un utilisateur",
                "operationId": "e75cbf2628721d46f0d8657a1e9e355b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Utilisateurs"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Utilisateurs"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "d6292122f5d9bd795ebef9ade111a1cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Utilisateur supprimé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/villes": {
            "get": {
                "tags": [
                    "Villes"
                ],
                "summary": "Lister villes",
                "operationId": "45da93baea60a6e346e095f1f1569e6b",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Villes"
                ],
                "summary": "Créer ville",
                "operationId": "1fa6a9eaa07969c49560b1d81c194187",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                }
            }
        },
        "/api/villes/{id}": {
            "get": {
                "tags": [
                    "Villes"
                ],
                "summary": "Afficher ville",
                "operationId": "7adf9a9bb69345c9df87412b4796f6e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails"
                    }
                }
            },
            "put": {
                "tags": [
                    "Villes"
                ],
                "summary": "Modifier ville",
                "operationId": "8958521a2fa599e035a7ae74b37ce61e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modifié"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Villes"
                ],
                "summary": "Supprimer ville",
                "operationId": "30fd610df817cb61b8cf58b20a179ad2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Utilisateurs": {
                "required": [
                    "nom",
                    "prenom",
                    "email",
                    "mot_de_passe",
                    "type_utilisateur_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "prenom": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mot_de_passe": {
                        "type": "string"
                    },
                    "type_utilisateur_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}